% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-emp-fdr.R, R/calc-emp-fdr-breaks.R,
%   R/plot-emp-fdr.R, R/plot-permuted-data.R
\name{calc_emp_fdr}
\alias{calc_emp_fdr}
\alias{calc_emp_fdr_breaks}
\alias{plot_emp_fdr}
\alias{plot_permuted_data}
\title{Calculate Empirical Number of False Positives}
\usage{
calc_emp_fdr(x, thresh_seq, warn = TRUE)

calc_emp_fdr_breaks(
  x,
  thresh_seq = seq(1, 0.1, by = -0.01),
  fdr_breaks = c(0.5, 1, 2, 3, 5)
)

plot_emp_fdr(x, ...)

plot_permuted_data(x, which = NULL, ...)
}
\arguments{
\item{x}{A \code{stab_sel} class object generated
via \code{\link[=stability_selection]{stability_selection()}}.}

\item{thresh_seq}{\code{numeric(n)}. A sequence in
\verb{[0, 1]} specifying the thresholds.
See \code{thresh} in \code{\link[=get_stable_features]{get_stable_features()}}}

\item{warn}{\code{logical(1)}. Should warnings be triggered if
mean of \verb{< 5} permutations is being returned?}

\item{fdr_breaks}{\code{numeric(n)}. A vector specifying the
desired mean number of empirical false positives at which
to calculate various thresholds.

\code{numeric(n)}. A vector specifying the}

\item{...}{Additional arguments passed to either
\code{\link[=calc_emp_fdr_breaks]{calc_emp_fdr_breaks()}} (i.e. \code{thresh_seq} and \code{fdr_breaks})
or \code{\link[=plot.stab_sel]{plot.stab_sel()}}.}

\item{which}{\code{integer(1)}. Which of the null hypothesis
permuted stability paths to plot.}
}
\value{
A named vector indicating the average number
(counts) of false positive features selected at the
various thresholds specified by \code{thresh_seq}.

For \code{\link[=calc_emp_fdr_breaks]{calc_emp_fdr_breaks()}}, a list containing:
\item{n_selected}{A vector of the number of features selected at each
empirical stability selection threshold}
\item{meanFPs}{A vector of the mean number of false positive selected
features at each empirical stability selection threshold}
\item{breaks}{A \code{tibble} of containing empirical false positive
summary statistics at each FDR specified break point}
}
\description{
Calculate the mean number of false positive features from
a permutation analysis performed during a stability selection
run. This assumes a permutation set was generated during
stability selection, (i.e. \code{n_perm > 0}).
}
\section{Functions}{
\itemize{
\item \code{calc_emp_fdr_breaks()}: Calculates the stability selection threshold, the mean number of
false positive selected features (empirical), and the number of
selected features for specified FDR break points.
Relies on \code{\link[=calc_emp_fdr]{calc_emp_fdr()}} to calculate the mean false discovery
based on permutations during the stability selection algorithm.

\item \code{plot_emp_fdr()}: plots the mean number of false positives (FPs)
versus the number of selected features by a sequence of
selection probability thresholds. For this to be possible,
the \code{stab_sel} object \emph{must} have permuted data in order to
calculate empirical false discovery rates. The area in the
sub-diagonal represents where more features are added without
a commensurate increase in false positives (\strong{Good}).
The inverse is true for the super-diagonal, false positives
are being included faster than additional features (\strong{Bad}).
The legend highlights pre-defined empirical FDR breaks:
\code{c(0.5, 1, 2, 3, 5)} evaluated to the nearest threshold cutoff.

\item \code{plot_permuted_data()}: Plot the permutation paths for a \code{stab_sel} object.
These paths are the stability selection paths
of the \code{n} class scrambled permutations, i.e. the null.

}}
\examples{
n_feat      <- 20
n_samples   <- 100
x           <- matrix(rnorm(n_feat * n_samples), n_samples, n_feat)
colnames(x) <- paste0("feat", "_", head(letters, n_feat))
y           <- sample(1:2, n_samples, replace = TRUE)

ss <- stability_selection(x, y, n_iter = 25, n_perm = 50,
                          r_seed = 101, parallel = TRUE)

calc_emp_fdr(ss, seq(0.5, 0.9, 0.1))

# calculate the FDR break points
calc_emp_fdr_breaks(ss)

# plot the FDR
plot_emp_fdr(ss)  # typically set permutations > 75

# Plot the permuted data individually
plot_permuted_data(ss, 3L)   # choose 3rd permutation
}
\seealso{
\code{\link[=stability_selection]{stability_selection()}}, \code{\link[=get_stable_features]{get_stable_features()}}
}
\author{
Stu Field, Michael R. Mehan
}
