% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmpFDR.R
\name{calcEmpFDR}
\alias{calcEmpFDR}
\title{Calculate Empirical Number of False Positives}
\usage{
calcEmpFDR(x, thresh.seq, warn = TRUE)
}
\arguments{
\item{x}{An object of class \code{stab_sel} generated via \code{\link[=stabilitySelection]{stabilitySelection()}}.}

\item{thresh.seq}{Numeric sequence in \verb{[0, 1]} specifying
the thresholds to evaluate.}

\item{warn}{Logical. Should warnings be triggered if
mean of \verb{< 5} permutations is being returned?}
}
\value{
A named vector indicating the average number (counts) of false
positive features selected at the various thresholds specified
by \code{thresh.seq}.
}
\description{
Calculate the mean number of false positive features from
a permutation analysis performed during a stability selection
run. This assumes a permutation set was generated during
stability selection, (i.e. \code{num.perms > 0}).
}
\examples{
withr::with_seed(101, {
  n_feat      <- 20
  n_samples   <- 100
  x           <- matrix(rnorm(n_feat * n_samples), n_samples, n_feat)
  colnames(x) <- paste0("feat", "_", head(letters, n_feat))
  y  <- sample(1:2, n_samples, replace = TRUE)
})
ss <- stabilitySelection(x, y, "l1-logistic", num.iter = 25,
                         num.perms = 25, r.seed = 101, parallel = TRUE)
calcEmpFDR(ss, seq(0.5, 0.9, 0.1))
}
\seealso{
\code{\link[=getThresholdFeatures]{getThresholdFeatures()}}

Other empirical FDR: 
\code{\link{calcEmpFDRbreaks}()},
\code{\link{plotEmpFDR}()}
}
\author{
Stu Field, Michael R. Mehan
}
\concept{empirical FDR}
