% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmpFDRbreaks.R
\name{calcEmpFDRbreaks}
\alias{calcEmpFDRbreaks}
\title{Calculate Empirical FDR Break Points}
\usage{
calcEmpFDRbreaks(
  x,
  thresh.seq = seq(1, 0.1, by = -0.01),
  fdr.breaks = c(0.5, 1, 2, 3, 5)
)
}
\arguments{
\item{x}{An object of class \code{stab_sel} generated via \code{\link[=stabilitySelection]{stabilitySelection()}}.}

\item{thresh.seq}{Numeric sequence in \verb{[0, 1]} specifying
the thresholds to evaluate.}

\item{fdr.breaks}{Numeric. A vector specifying the desired mean number
of empirical false positives at which to calculate various thresholds.}
}
\value{
A list consisting of:
\item{n_selected}{A vector of the number of features selected at each
empirical stability selection threshold}
\item{meanFPs}{A vector of the mean number of false positive selected
features at each empirical stability selection threshold}
\item{breaks}{A \code{tibble} of containing empirical false positive
summary statistics at each FDR specified break point}
}
\description{
Calculates the stability selection threshold, the mean number of
false positive selected features (empirical), and the number of
selected features for specified FDR break points.
Relies on \code{\link[=calcEmpFDR]{calcEmpFDR()}} to calculate the mean false discovery
based on permutations during the stability selection algorithm.
}
\examples{
# l1-logistic
withr::with_seed(101, {
  n_feat      <- 20
  n_samples   <- 100
  x           <- matrix(rnorm(n_feat * n_samples), n_samples, n_feat)
  colnames(x) <- paste0("feat", "_", head(letters, n_feat))
  y        <- sample(1:2, n_samples, replace = TRUE)
})
stab_sel <- stabilitySelection(x, y, "l1-logistic", num.iter = 25,
                               num.perms = 25,
                               r.seed = 101, parallel = TRUE)
calcEmpFDRbreaks(stab_sel)
}
\seealso{
\code{\link[=getStableFeatures]{getStableFeatures()}}, \code{\link[=stabilitySelection]{stabilitySelection()}}

Other empirical FDR: 
\code{\link{calcEmpFDR}()},
\code{\link{plotEmpFDR}()}
}
\author{
Stu Field
}
\concept{empirical FDR}
