% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stable-features.R
\name{get_stable_features}
\alias{get_stable_features}
\title{Calculate Stable Features}
\usage{
get_stable_features(x, thresh, add_features, warn)
}
\arguments{
\item{x}{An \code{stab_sel} class object OR a matrix containing
selection probabilities, i.e. the \code{stabpath_matrix} entry
of a \code{stab_sel} object.}

\item{thresh}{\code{numeric(1)} in \verb{[0, 1]}. the minimum
selection probability threshold.
In some instances this value can also be a vector, but is
generally a scalar > 0.50 for \code{\link[=get_stable_features]{get_stable_features()}}.}

\item{add_features}{\code{character(n)}. A string of additional features
to \emph{force} into the resulting table, irrespective of their
threshold. Used mostly in the S3 plot method to see a given
stability path of a feature not meeting a threshold cutoff.
Must be exact string match.}

\item{warn}{\code{logical(1)}. Should warnings be triggered if no stable
features were found at the specified threshold OR if the
FDR upper bound is undefined at thresholds \verb{<= 0.5}?}
}
\value{
A two column data frame containing maximum selection
probabilities and FDR upper bounds when appropriate, see \code{Details}.
}
\description{
Returns a data frame object of all features with a maximum
selection probability greater than a minimum threshold.
}
\details{
A "stable feature" is defined as a feature with a maximum selection
probability greater than a supplied threshold. This function returns a
\code{data.frame} of all features that satisfy this criterion along with the
maximum selection probability and the upper bound on the false discover
rate. This false discovery rate bound is only defined for \code{thresh > 0.5},
it is otherwise undefined.

\emph{IMPORTANT!} If you pass to the \code{matrix} method, there
is no permutation analysis performed, i.e. no \verb{$EmpFDR} column
in the returned data frame. This calculation takes a long
time and is not always desired, so this method offers the
user a control mechanism for the output behavior.
}
\examples{
# logistic regression
n_feat      <- 20
n_samples   <- 100
x           <- matrix(rnorm(n_feat * n_samples), n_samples, n_feat)
colnames(x) <- paste0("feat", "_", head(letters, n_feat))
y           <- sample(1:2, n_samples, replace = TRUE)

stab_sel <- stability_selection(x, y, r_seed = 101)

# Stable features at `thresh =`
get_stable_features(stab_sel, 0.75)

get_stable_features(stab_sel, c(0.75, 0.9))
}
\seealso{
\code{\link[=stability_selection]{stability_selection()}}
}
\author{
Stu Field
}
