% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-emp-fdr.R, R/plot-permuted-data.R
\name{plot_emp_fdr}
\alias{plot_emp_fdr}
\alias{plot_permuted_data}
\title{Plot Empirical FDR}
\usage{
plot_emp_fdr(x, thresh_seq = seq(1, 0.1, by = -0.01))

plot_permuted_data(x, which = NULL, ...)
}
\arguments{
\item{x}{A \code{stab_sel} class object generated
via \code{\link[=stability_selection]{stability_selection()}}.}

\item{thresh_seq}{\code{numeric(n)}. A sequence in
\verb{[0, 1]} specifying the thresholds to evaluate.}

\item{which}{\code{integer(1)}. Which of the null hypothesis
permuted stability paths to plot.}

\item{...}{Additional arguments passed to \code{\link[=plot.stab_sel]{plot.stab_sel()}}.}
}
\description{
Plot the mean number of false positives (FPs) versus the number of selected
features by a sequence of selection probability thresholds. For this to
be possible, the \code{stab_sel} object \emph{must} have permuted
data in order to calculate empirical false discovery rates. The area
in the sub-diagonal represents where more features are added without
a commensurate increase in false positives (\strong{Good}). The inverse is true
for the super-diagonal, false positives are being included faster
than additional features (\strong{Bad}). The legend highlights pre-defined
empirical FDR breaks: \code{c(0.5, 1, 2, 3, 5)} evaluated to the nearest
threshold cutoff.
}
\section{Functions}{
\itemize{
\item \code{plot_permuted_data()}: Plot the permutation paths for an object of
class \code{stab_sel}.
These paths are the stability selection paths
of the \code{n} class scrambled permutations, i.e. the null.

}}
\examples{
# l1-logistic
withr::with_seed(101, {
  n_feat      <- 20
  n_samples   <- 100
  x           <- matrix(rnorm(n_feat * n_samples), n_samples, n_feat)
  colnames(x) <- paste0("feat", "_", head(letters, n_feat))
  y           <- sample(1:2, n_samples, replace = TRUE)
})

# typically set > 75 permutations
stab_sel <- stability_selection(x, y, "l1-logistic", num_perms = 25,
                                r_seed = 101, parallel = TRUE)
plot_emp_fdr(stab_sel)

# Plot the permuted data individually
plot_permuted_data(stab_sel, 3L)   # choose 3rd permutation
}
\seealso{
\code{\link[=stability_selection]{stability_selection()}}, \code{\link[=get_stable_features]{get_stable_features()}}

Other empirical FDR: 
\code{\link{calc_emp_fdr}()},
\code{\link{calc_emp_fdr_breaks}()}
}
\author{
Stu Field
}
\concept{empirical FDR}
