% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmpFDR.R, R/plotPermutedData.R
\name{plotEmpFDR}
\alias{plotEmpFDR}
\alias{plotPermutedData}
\title{Plot Empirical FDR}
\usage{
plotEmpFDR(x, thresh.seq = seq(1, 0.1, by = -0.01))

plotPermutedData(x, which = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{stab_sel} generated via \code{\link[=stabilitySelection]{stabilitySelection()}}.}

\item{thresh.seq}{Numeric sequence in \verb{[0, 1]} specifying
the thresholds to evaluate.}

\item{which}{Integer. Indicating which of the null hypothesis
permuted stability paths to plot.}

\item{...}{Additional arguments passed to \code{\link[=plot.stab_sel]{plot.stab_sel()}}.}
}
\description{
Plot the mean number of false positives (FPs) versus the number of selected
features by a sequence of selection probability thresholds. For this to
be possible, the \code{stab_sel} object \emph{must} have permuted
data in order to calculate empirical false discovery rates. The area
in the sub-diagonal represents where more features are added without
a commensurate increase in false positives (\strong{Good}). The inverse is true
for the super-diagonal, false positives are being included faster
than additional features (\strong{Bad}). The legend highlights pre-defined
empirical FDR breaks: \code{c(0.5, 1, 2, 3, 5)} evaluated to the nearest
threshold cutoff.
}
\section{Functions}{
\itemize{
\item \code{plotPermutedData()}: Plot the permutation paths for an object of class \code{stab_sel}.
These paths are the stability selection paths of the \code{n} class scrambled
permutations, i.e. the null.

}}
\examples{
# l1-logistic
withr::with_seed(101, {
  n_feat      <- 20
  n_samples   <- 100
  x           <- matrix(rnorm(n_feat * n_samples), n_samples, n_feat)
  colnames(x) <- paste0("feat", "_", head(letters, n_feat))
  y           <- sample(1:2, n_samples, replace = TRUE)
})

# typically set > 75 permutations
stab_sel <- stabilitySelection(x, y, "l1-logistic", num.perms = 25,
                               r.seed = 101, parallel = TRUE)
plotEmpFDR(stab_sel)

# Plot the permuted data individually
plotPermutedData(stab_sel, 3L)   # choose 3rd permutation
}
\seealso{
\code{\link[=stabilitySelection]{stabilitySelection()}}, \code{\link[=getStableFeatures]{getStableFeatures()}}

Other empirical FDR: 
\code{\link{calcEmpFDR}()},
\code{\link{calcEmpFDRbreaks}()}
}
\author{
Stu Field
}
\concept{empirical FDR}
