% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny-cluster.R, R/s3-plot-pclust.R,
%   R/s3-print-pclust.R
\name{progeny_cluster}
\alias{progeny_cluster}
\alias{is_pclust}
\alias{plot.pclust}
\alias{print.pclust}
\title{Perform Progeny Clustering}
\usage{
progeny_cluster(data, clust_iter = 2:10L, repeats = 10L, r_seed = 1234, ...)

is_pclust(x)

\method{plot}{pclust}(x, ...)

\method{print}{pclust}(x, ...)
}
\arguments{
\item{data}{A \eqn{n \times p} data matrix containing \emph{n} samples
and \emph{p} features. Can also be a data frame where each row
corresponds to a sample or observation, and each column
corresponds to a feature or variable.}

\item{clust_iter}{\code{integer(n)}. Span of \code{k} clusters to
interrogate over.}

\item{repeats}{\code{integer(1)}. The number of repeat iterations to
perform. Particularly useful if error bars during plotting
are desired.}

\item{r_seed}{\code{integer(1)}. Seed for the random number
generator, for reproducibility.}

\item{...}{Important! Parameters passed to the internal
\code{progeny_k()}, i.e. \verb{n_iter =} and \verb{size =}. Or,
for the \code{\link[=plot]{plot()}} method, arguments passed to the corresponding
graphics device.}

\item{x}{A \code{pclust} class object (or an object to be tested for one).}
}
\value{
A \code{pclust} class object, a list containing:
\item{scores}{A matrix of stability scores for each iteration in a matrix,
with \code{k} columns}
\item{mean_scores}{The mean stability scores for each cluster \code{k}}
\item{ci95_scores}{95\% confidence interval scores}
\item{random_scores}{The reference (random) scores for each
iteration at each clustering level (\code{k})}
\item{mean_random_scores}{The mean of the reference (random) data set, i.e.
column means of \code{random_scores}}
\item{D_max}{The distance between the mean stability scores and the mean
reference scores for each cluster \code{k}}
\item{D_gap}{The "gap" distance metric for neighboring cluster k differences.
See original paper for reference.}
\item{clust_iter}{Integer Sequence of \code{k} clusters interrogated}
\item{repeats}{The number of repeat iterations to performed}
\item{n_iter}{The number of progeny iterations to performed}
\item{size}{The progeny size used in each iteration}
\item{call}{The call made to \code{\link[=progeny_cluster]{progeny_cluster()}}}

\code{\link[=is_pclust]{is_pclust()}} returns a logical boolean.
}
\description{
Determine the most stable (optimal) number of clusters via
Progeny Clustering algorithm.

The \code{\link[=is_pclust]{is_pclust()}} function checks whether
an object is class \code{pclust}. See \code{\link[=inherits]{inherits()}}.
}
\examples{
# `n_iter =` and `size =` are passed to `progeny_k()`
pclust <- progeny_cluster(progeny_data, clust_iter = 2:9L,
                          n_iter = 20L, size = 6)
pclust

# Test progeny clustering on iris data set
# Doesn't work as well as the simulated data set
clust_iris <- progeny_cluster(iris[, -5L], clust_iter = 2:5L,
                              size = 6L, n_iter = 250)  # pass `...`

# true n_clusters = 3
clust_iris

# Test for class `pclust`
is_pclust(pclust)

# S3 plot method
plot(pclust)

plot(clust_iris)
}
\references{
Hu, CW, Kornblau, SM, Slater, JH and AA Qutub (2015).
Progeny Clustering: A Method to Identify Biological Phenotypes.
Scientific Reports, \strong{5}:12894. \url{http://www.nature.com/articles/srep12894}
}
\seealso{
\code{\link[stats:kmeans]{stats::kmeans()}}

Other cluster: 
\code{\link{stability_cluster}()}
}
\author{
Stu Field
}
\concept{cluster}
