% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabilityselectr-package.R
\docType{package}
\name{stabilityselectr-package}
\alias{stabilityselectr}
\alias{stabilityselectr-package}
\title{stabilityselectr: R Package to Perform Stability and Cluster Selection}
\description{
The Stability Selection package performs stability selection with a variety of kernels provided by the 'glmnet' package, and provides simple tools for plotting and extracting selected features. There is additional functionality designed to facilitate various forms of permutation clustering analyses.
}
\details{
Stability selection is performed using \code{\link[=stabilitySelection]{stabilitySelection()}},
which returns a \code{stab_sel} class object. The stability path can be
plotted using \code{\link[graphics:plot.default]{graphics::plot()}}. A \code{tibble} with the highest
selection probability can be created using \code{\link[=getStableFeatures]{getStableFeatures()}}.
See vignettes.
}
\references{
\strong{Meinshausen, N. and Buhlmann, P.} (\emph{2010}), Stability selection.
Journal of the Royal Statistical Society: Series B (Statistical
Methodology), 72: 417-473. doi: 10.1111/j.1467-9868.2010.00740.x

\strong{Hu, C.W., Kornblau, S.M., Slater, J.H. and A.A. Qutub} (\emph{2015}).
Progeny Clustering: A Method to Identify Biological Phenotypes. Scientific Reports,
5:12894. \url{http://www.nature.com/articles/srep12894}
}
\seealso{
\code{\link[glmnet:glmnet]{glmnet::glmnet()}}, \code{\link[=kmeans]{kmeans()}}
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
\keyword{package}
