% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-emp-fdr.R
\name{calc_emp_fdr}
\alias{calc_emp_fdr}
\title{Calculate Empirical Number of False Positives}
\usage{
calc_emp_fdr(x, thresh_seq, warn = TRUE)
}
\arguments{
\item{x}{A \code{stab_sel} class object generated
via \code{\link[=stability_selection]{stability_selection()}}.}

\item{thresh_seq}{\code{numeric(n)}. A sequence in
\verb{[0, 1]} specifying the thresholds to evaluate.}

\item{warn}{\code{logical(1)}. Should warnings be triggered if
mean of \verb{< 5} permutations is being returned?}
}
\value{
A named vector indicating the average number
(counts) of false positive features selected at the
various thresholds specified by \code{thresh_seq}.
}
\description{
Calculate the mean number of false positive features from
a permutation analysis performed during a stability selection
run. This assumes a permutation set was generated during
stability selection, (i.e. \code{num_perms > 0}).
}
\examples{
withr::with_seed(101, {
  n_feat      <- 20
  n_samples   <- 100
  x           <- matrix(rnorm(n_feat * n_samples), n_samples, n_feat)
  colnames(x) <- paste0("feat", "_", head(letters, n_feat))
  y  <- sample(1:2, n_samples, replace = TRUE)
})
ss <- stability_selection(x, y, "l1-logistic", num_iter = 25,
                          num_perms = 25, r_seed = 101, parallel = TRUE)
calc_emp_fdr(ss, seq(0.5, 0.9, 0.1))
}
\seealso{
\code{\link[=get_threshold_features]{get_threshold_features()}}

Other empirical FDR: 
\code{\link{calc_emp_fdr_breaks}()},
\code{\link{plot_emp_fdr}()}
}
\author{
Stu Field, Michael R. Mehan
}
\concept{empirical FDR}
