% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-stable-features.R,
%   R/get-threshold-features.R
\name{get_stable_features}
\alias{get_stable_features}
\alias{get_threshold_features}
\title{Calculate Stable Features}
\usage{
get_stable_features(x, thresh, add_features, warn)

get_threshold_features(x, thresh_vec, ...)
}
\arguments{
\item{x}{An object of class \code{stab_sel} OR a matrix containing
selection probabilities, i.e. the \code{stabpath.matrix} entry
of a \code{stab_sel} class object.}

\item{thresh}{Numeric in \verb{[0, 1]}. Minimum selection
probability threshold.}

\item{warn}{\code{logical(1)}. Should warnings be triggered if no stable
features were found at the specified threshold OR if the
FDR upper bound is undefined at thresholds \verb{<= 0.5}?}

\item{thresh_vec}{\code{numeric(1)}. A vector of threshold values.}

\item{...}{Additional arguments passed to \code{\link[=get_stable_features]{get_stable_features()}},
typically the \verb{add_features =} and \verb{warn =} arguments.}

\item{add.features}{Character. A string of additional features
to \emph{force} into the resulting table, irrespective of their
threshold. Used mostly in the S3 plot method to see a given
stability path of a feature not meeting a threshold cutoff.
Must be exact string match.}
}
\value{
A two column data frame containing maximum selection probabilities
and FDR upper bounds when appropriate, see \code{Details}.

For \code{\link[=get_threshold_features]{get_threshold_features()}}, a \emph{list} of
data frames for various thresholds corresponding
to the entries of \code{thresh_vec}.
}
\description{
Returns a data frame object of all features with a maximum
selection probability greater than a minimum threshold.

\code{\link[=get_threshold_features]{get_threshold_features()}} calculates the features at or above a given
selection probability threshold. It is a thin wrapper around
\code{\link[=get_stable_features]{get_stable_features()}} returning a list of thresholds corresponding
to \code{thresh_vec}.
}
\details{
A "stable feature" is defined as a feature with a maximum selection
probability greater than a supplied threshold. This function returns a
\code{data.frame} of all features that satisfy this criterion along with the
maximum selection probability and the upper bound on the false discover
rate. This false discovery rate bound is only defined for \code{thresh > 0.5},
it is otherwise undefined.

\emph{IMPORTANT!} If you pass to the \code{matrix} method, there
is no permutation analysis performed, i.e. no \verb{$EmpFDR} column
in the returned data frame. This calculation takes a long
time and is not always desired, so this method offers the
user a control mechanism for the output behavior.
}
\examples{
# l1-logistic
withr::with_seed(101, {
  n_feat      <- 20
  n_samples   <- 100
  x           <- matrix(rnorm(n_feat * n_samples), n_samples, n_feat)
  colnames(x) <- paste0("feat", "_", head(letters, n_feat))
  y           <- sample(1:2, n_samples, replace = TRUE)
})

stab_sel <- stability_selection(x, y)

# Stable features at thresh = 0.55
get_stable_features(stab_sel, 0.55)
# l1-logistic
withr::with_seed(101, {
  n_feat      <- 20
  n_samples   <- 100
  x           <- matrix(rnorm(n_feat * n_samples), n_samples, n_feat)
  colnames(x) <- paste0("feat", "_", head(letters, n_feat))
  y           <- sample(1:2, n_samples, replace = TRUE)
})
stab_sel <- stability_selection(x, y, "l1-logistic", r_seed = 101)
get_threshold_features(stab_sel, seq(0.6, 0.9, 0.1))
}
\seealso{
\code{\link[=stability_selection]{stability_selection()}}
}
\author{
Stu Field
}
