% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny-cluster.R, R/s3-plot-pclust.R,
%   R/s3-print-pclust.R
\name{progeny_cluster}
\alias{progeny_cluster}
\alias{is_pclust}
\alias{plot.pclust}
\alias{print.pclust}
\title{Perform Progeny Clustering}
\usage{
progeny_cluster(
  data,
  clust_iter = 2:10L,
  reps = 10L,
  verbose = interactive(),
  ...
)

is_pclust(x)

\method{plot}{pclust}(x, ..., file = NULL)

\method{print}{pclust}(x, ...)
}
\arguments{
\item{data}{A (\verb{n x p}) data matrix containing \emph{n} samples and \emph{p} features.
Can also be a data frame where each row corresponds to a sample or
observation, whereas each column corresponds to a feature or variable.}

\item{clust_iter}{\code{integer(n)}. Span of \code{k} clusters to interrogate}

\item{reps}{\code{integer(1)}. The number of repeat iterations to perform.
Particularly useful if error bars during plotting are desired.}

\item{verbose}{\code{logical(1)}. Print the progress of the clustering repeats
to the console. Defaults to \code{\link[=interactive]{interactive()}}.}

\item{...}{Additional parameters passed to the internal \code{progenyK()},
typically \verb{iter =} and \verb{size =}.}

\item{x}{A \code{pclust} object.}

\item{file}{\code{character(1)}. A file name and
path to save the plot to an external file.}
}
\value{
An object of class \code{pclust}, a list containing:
\item{scores:}{A matrix of stability scores for each iteration in a matrix,
with \code{k} columns}
\item{mean_scores:}{The mean stability scores for each cluster \code{k}}
\item{ci95_scores:}{95\% confidence interval scores}
\item{random_scores:}{The reference (random) scores for each iteration
at each clustering level (\code{k})}
\item{mean_random_scores:}{The mean of the reference (random) data set, i.e.
column means of \code{random.scores}}
\item{D_max:}{The distance between the mean stability scores and the mean
reference scores for each cluster \code{k}}
\item{D_gap:}{The "gap" distance metric for neighboring cluster k differences.
See original paper for reference.}
\item{clust_iter:}{Integer Sequence of \code{k} clusters interrogated}
\item{repeats:}{The number of repeat iterations to performed}
\item{iter:}{The number of progeny iterations to performed}
\item{size:}{The progeny size used in each iteration}
\item{call:}{The call made to \code{\link[=progeny_cluster]{progeny_cluster()}}}

\code{\link[=is_pclust]{is_pclust()}} returns a logical boolean.
}
\description{
Determine the most stable (optimal) number of clusters via
Progeny Clustering algorithm.

The \code{\link[=is_pclust]{is_pclust()}} function checks whether
an object is of class \code{pclust}. See \code{\link[=inherits]{inherits()}}.
}
\examples{
# `iter=` and `size=` are passed to `progeny_k()`
pclust <- withr::with_seed(1234,
  progeny_cluster(progeny_data, clust_iter = 2:9L, iter = 20L, size = 6)
)

# Test progeny clustering on Iris data set
# Doesn't work quite as well as the simulated data set
clustIris <- withr::with_seed(99,
  progeny_cluster(iris[, -5L], clust_iter = 2:5L, size = 6L, iter = 50)
)
clustIris    # true n clusters = 3

# Test for class `pclust`
is_pclust(pclust)

# S3 plot method
plot(pclust)
plot(clustIris)

# S3 print method
pclust
clustIris

}
\references{
Hu, C.W., Kornblau, S.M., Slater, J.H. and A.A. Qutub (2015).
Progeny Clustering: A Method to Identify Biological Phenotypes.
Scientific Reports, 5:12894. \url{http://www.nature.com/articles/srep12894}
}
\seealso{
\code{\link[stats:kmeans]{stats::kmeans()}}

Other cluster: 
\code{\link{stability_cluster}()}
}
\author{
Stu Field
}
\concept{cluster}
